Conversores

Description

Este módulo contem várias funções de conversão usadas em vários módulos do programa.

Documentation

getLargura :: Jogo -> Largura

Função para determinar a largura de um mapa.

agarraJogador :: Jogo -> Jogador

Função para determinar o jogador de um jogo.

agarraMapa :: Jogo -> Mapa

Função para determinar o mapa de um jogo.

desconstroiMapa :: Mapa -> [[Obstaculo]]

Função para converter um mapa numa matriz de obstaculos.

agarraObstaculos :: Jogo -> [[Obstaculo]]

Função para converter um Jogo numa matriz de obstaculos.

obstaculoToString :: [Obstaculo] -> [StringObstaculo]

Função para converter um obstaculo para uma string.

agarraTerrenos :: Jogo -> [[Terreno]]

Função para converter um Jogo numa matriz de Terrenos.

terrenoToString :: [Terreno] -> [StringTerreno]

Função para converter um terreno para uma string.

terrenoToString2 :: Terreno -> StringTerreno

Função para converter um terreno para uma string.

conversorModelopersonagemToNome :: Modelopersonagem -> String

Função para converter um Modelo de personagem para uma string.

conversorModeloMapaToInt :: ModeloMapa -> Int

Função para converter um Modelo de mapa para um inteiro.

conversorRoupasToString :: Roupas -> String

Função para converter o tipo Roupas para uma string.