Main

Description

Este módulo contem a resolução da Tarefa 6, a qual contém a aplicação gráfica completa.

O grupo apresenta assim o trabalho concluído, com plena noção de falhas exporádicas em algumas features requeridas e outras não.

É de notar que o modo Carreira não cumpre, propositadamente, as regras presentes no guião.

O trabalho apresentava uma potencialidade de desenvolvimento e aperfeiçoamento enorme, no entanto, tal não foi possível devido aos prazos estabelecidos e a cumprir.

Bibliografia: -- ** Guião Gloss feito pelo CeSIUM - https://www.youtube.com/watch?v=VDVGO3BTGYk&list=PLadvWyx_6w6XiJ95A4MqSfmIaRVbXWFGS -- ** Documentação do Gloss - https://hackage.haskell.org/package/gloss -- ** Fórum Stack Overflow -- ** Material disponibilizado pelos docentes da UC

Documentation

type EstadoGloss = (Menu, [Jogo], TexturasTerrenos, TexturasObstaculos, Extras, TexturasMenu, Int, ControladorBot, Modelopersonagem, ModeloMapa, Skins, [Roupas], TopScore, TempoDecorrido, ControladorSingle, ControladorMulti, [Score])

O estado do programa.

l :: Float

Distância entre duas pictures no mapa.

altura :: Float

Altura do mapa.

comprimento :: Float

Comprimento do mapa.

frames :: Int

Uma constante que define a taxa de atualização do programa - quantas vezes a função reageTempoGloss é chamada por segundo.

dimensions :: Display

Define a dimensão da janela (neste caso FullScreen).

desenhaTerreno :: Float -> Float -> ModeloMapa -> StringTerreno -> TexturasTerrenos -> Picture

Função que desenha um terreno.

desenhaLinhaTerreno :: Float -> Float -> ModeloMapa -> [StringTerreno] -> TexturasTerrenos -> [Picture]

Função que desenha uma linha de terrenos.

desenhaMapaTerreno :: Float -> Float -> ModeloMapa -> [[StringTerreno]] -> TexturasTerrenos -> [Picture]

Função que desenha um mapa de terrenos.

desenhaObstaculo :: Float -> Float -> ModeloMapa -> StringObstaculo -> TexturasObstaculos -> Float -> Picture

Função que desenha um obstaculo.

desenhaLinhaObstaculo :: Float -> Float -> ModeloMapa -> [StringObstaculo] -> TexturasObstaculos -> Float -> [Picture]

Função que desenha uma linha de obstaculos.

desenhaMapaObstaculo :: Float -> Float -> ModeloMapa -> [[StringObstaculo]] -> TexturasObstaculos -> Float -> [Picture]

Função que desenha um mapa de obstaculos.

desenhaMenu :: StringMenu -> TexturasMenu -> Picture

Função que desenha um menu.

desenhaTopBar :: Picture

Função que desenha uma topbar para o jogo.

corMoldura :: Color

Função que define a cor da moldura.

desenhaMoldura :: String -> [Picture]

Função que desenha a moldura do jogo.

desenhaMolduraLoja :: String -> [Picture]

Função que desenha a moldura na Loja.

desenhaScore :: StringExtras -> Extras -> Picture

Função que desenha um valor de score.

desenhaJogador :: ModeloMapa -> Jogador -> Roupas -> Skins -> Picture

Função que desenha um jogador.

realPlayerX :: Int -> Float

Função auxiliar à desenhaJogador para converter as coordenadas à escala do mapa.

realPlayerY :: Int -> Float

Função auxiliar à desenhaJogador para converter as coordenadas à escala do mapa.

desenhaJogadorLoja :: ModeloMapa -> Jogador -> Roupas -> Skins -> Picture

Função que desenha um jogador na Loja.

realPlayerXLoja :: Int -> Float

Função auxiliar à desenhaJogador para converter as coordenadas à escala do mapa.

realPlayerYLoja :: Int -> Float

Função auxiliar à desenhaJogador para converter as coordenadas à escala do mapa.

estadoInicial :: [Jogo]

Estado incial do programa.

estadoGlossInicial :: Int -> Extras -> [Roupas] -> TopScore -> TexturasTerrenos -> TexturasObstaculos -> TexturasMenu -> Skins -> EstadoGloss

Função que define o estado inicial do programa.

desenhaEstadoGloss :: EstadoGloss -> IO Picture

Desenha o estado do programa, consoante algumas variáveis - transforma um estado numa Picture.

drawOption :: String -> Picture

Função que define padrão de escrita para a topbar.

reageEventoGloss :: Event -> EstadoGloss -> IO EstadoGloss

Função que reage a um evento, por parte do utilizador, através do teclado ou do rato. Reagir Menu

reageTempoGloss :: Float -> EstadoGloss -> IO EstadoGloss

Função que reage à passagem do tempo. Reagir ao longo do tempo

tiraUMAlinha :: Mapa -> Mapa

Função auxiliar ao desliza do mapa.

lerFicheiro :: IO ([Roupas], TopScore)

Função que permite leitura de informação de um ficheiro.

main :: IO ()

Função principal.

u :: Velocidade

Exemplos de jogo

jogo1 :: Jogo

jogo2 :: Jogo