Tarefa3_2022li1g010

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Função animaJogo

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo recebe: Jogo -> Jogada e devolve Jogo. Esta função recorre às funções auxiliares posJogadorparado e postroncoaux, para no fim receber o tipo esperado (Jogo).

animaJogo :: Jogo -> Jogada -> Jogo
animaJogo (Jogo jogador mapa) movimento = (Jogo (posJogadorparado (Jogo jogador mapa) movimento) (postroncoaux mapa))

Exemplo de utilização:

>>> animaJogo (Jogo (Jogador (1,2)) (Mapa 5 [(Rio 1, [Nenhum,Tronco,Tronco,Nenhum,Nenhum]),(Rio 1, [Tronco,Tronco,Nenhum,Nenhum,Tronco]),(Estrada 2,[Carro,Nenhum,Carro,Carro,Nenhum])])) (Move Direita)] = Jogo (Jogador (2,2)) (Mapa 5 [(Rio 1,[Nenhum,Nenhum,Tronco,Tronco,Nenhum]),(Rio 1,[Tronco,Tronco,Tronco,Nenhum,Nenhum]),(Estrada 2,[Carro,Nenhum,Carro,Nenhum,Carro])])

Propriedades:

Esta função não determina a validade de um jogador fora de um mapa, limita-se apenas na permitibilidade de movimentos do mesmo.

Funções auxiliares

Função postroncoaux

postroncoaux :: Mapa -> Mapa

A função postroncoaux recebe um Mapa e dá como resultado um Mapa. O objetivo desta função é converter o resultado da função posTroncoECarro para um Mapa, para poder ser utilizado na função animaJogo.

postroncoaux :: Mapa -> Mapa  
postroncoaux (Mapa largura ((terreno,l):ts)) = (Mapa largura (posTroncoECarro ((terreno,l):ts)))

Exemplo de utilização:

>>> postroncoaux (Mapa 4 [(Rio 1,[Nenhum,Tronco,Tronco,Nenhum]),(Estrada 2, [Nenhum,Carro,Carro,Nenhum]), (Relva, [Nenhum,Arvore,Arvore,Nenhum])]) = Mapa 4 [(Rio 1,[Nenhum,Nenhum,Tronco,Tronco]),(Estrada 2,[Carro,Nenhum,Nenhum,Carro]),(Relva,[Nenhum,Arvore,Arvore,Nenhum])]
>>> postroncoaux (Mapa 4 [(Rio 1,[Nenhum,Tronco,Tronco,Nenhum]),(Estrada (-2), [Nenhum,Carro,Carro,Nenhum]), (Relva, [Nenhum,Arvore,Arvore,Nenhum])]) = Mapa 4 [(Rio 1,[Nenhum,Nenhum,Tronco,Tronco]),(Estrada (-2),[Carro,Nenhum,Nenhum,Carro]),(Relva,[Nenhum,Arvore,Arvore,Nenhum])]

Função posTroncoECarro

posTroncoECarro :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

A função posTroncoECarro recebe uma lista de tuplos do tipo [(Terreno,[Obstaculo])] e dá como resultado uma outra lista do mesmo tipo. Como sabemos os terrenos : Estrada e Rio têm uma determinada velocidade, logo nesta função os obstáculos que fazem parte destes terrenos irão movimentar-se um número de vezes de acordo com a sua velocidade.

Desta forma a função posTroncoECarro, chama recursivamente a função posaux, até ficar com uma lista vazia, nesse momento a recursividade pára.

posTroncoECarro :: [(Terreno,[Obstaculo])] -> [(Terreno,[Obstaculo])]
posTroncoECarro [] = []
posTroncoECarro ((Relva, l):ts) = (Relva, l) : posTroncoECarro ts 
posTroncoECarro ((Rio a, l):ts) = (Rio a, (posaux l a)) :posTroncoECarro ts
posTroncoECarro ((Estrada a, l):ts) = (Estrada a, (posaux l a)) : posTroncoECarro ts 

Propriedades:

O tipo de terreno : __Relva__ não tem velocidade, logo os seus obstáculos são fixos.

Função posaux

posaux :: [Obstaculo] -> Int -> [Obstaculo]

A função posaux recebe uma lista dos tipos Obstaculos e ainda recebe um Int que corresponde à velocidade de um determinado terreno. Esta função é a encarregada de, para cada terreno diferente, que sejam feitas as movimentações dos seus obstáculos.

posaux :: [Obstaculo] -> Int -> [Obstaculo]
posaux l 0 = l 
posaux l a 
    | a > 0 = posaux (last l: init l) (a-1)
    | a < 0 = posaux (tail l ++ [head l]) (a+1)

Propriedades:

A função para a sua recursividade quando o Int for 0.

Se a velocidade é maior que 0, logo o terreno tem direção horizontal e sentido de esquerda a direita.

Se a velocidade é menor que 0, logo o terreno tem direção horizontal e sentido de direita a esquerda.

porp> Se a velocidade é 0, isto sinifica que os objetos não mudam mais de posição.

Função posJogadorparado

posJogadorparado :: Jogo -> Jogada -> Jogador

A função posJogadorparado recebe um Jogo e uma Jogada e dá como resultado um Jogador. O objetivo desta função é efetuar a Jogada inserida na função animaJogo. Esta função reccore a 3 funções auxiliares : posjogador, localizajogador e localizavelocidade. Desta forma, a função considerará para qualquer que seja o terreno, as diferentes possibilidades de Jogada, respeitando as diferentes regras básicas do jogo.

posJogadorparado :: Jogo -> Jogada -> Jogador
posJogadorparado jogo(Jogo (Jogador (x,y)) (Mapa largura linhas)) Parado = if ((localizajogador jogo) !! x == Tronco)
                                                                            then (Jogador ((x + localizavelocidade jogo),y))
                                                                            else posjogador jogo Parado

posJogadorparado jogo(Jogo (Jogador (x,y)) (Mapa largura linhas)) (Move Cima) = if ((localizajogador jogo) !! x == Tronco) && ((snd(linhas !! (y-1))) !! x) == Arvore
                                                                                 then (Jogador ((x + localizavelocidade jogo),y)) 
                                                                                 else posjogador jogo (Move Cima) 

posJogadorparado jogo(Jogo (Jogador (x,y)) (Mapa largura linhas)) (Move Baixo) = if ((localizajogador jogo) !! x == Tronco) && ((snd(linhas !! (y+1))) !! x) == Arvore
                                                                                  then (Jogador ((x + localizavelocidade jogo),y))  
                                                                                  else posjogador jogo (Move Baixo) 

posJogadorparado jogo(Jogo (Jogador (x,y)) (Mapa largura linhas)) (Move Direita) = if ((localizajogador jogo) !! x == Tronco) 
                                                                                    then (Jogador ((x + 1 + localizavelocidade jogo),y)) 
                                                                                    else posjogador jogo (Move Direita) 
                                                                                    
posJogadorparado jogo@(Jogo (Jogador (x,y)) (Mapa largura linhas)) (Move Esquerda) = if ((localizajogador jogo) !! x == Tronco) 
                                                                                     then (Jogador ((x - 1 + localizavelocidade jogo),y))  
                                                                                     else posjogador jogo (Move Esquerda) 

Propriedades:

O jogador nunca pode ir para cima de uma /arvore/.

Se o jogador se encontra num tronco, este deverá movimentar-se com a velocidade do mesmo.

Se o jogador efetua uma  Jogada : __Cima__ ou __Baixo__ e ao mesmo tempo se encontrar num tronco, este não adquire a velocidade do tronco e efetua a __Jogada__ correspondente, caso o jogador tenha à frente ou atrás uma Arvore, matem-se no tronco e movimenta-se com a velocidade deste.

Função posjogador

posjogador :: Jogo -> Jogada -> Jogador

A função posjogador recebe um Jogo e uma Jogada e dá como resultado um Jogador. O objetivo desta função é alterar as coordenadas do Jogador segundo o tipo de Jogada que é inserido. Além disto, esta função tem em conta os limites do mapa e a condicionante Arvore.

posjogador :: Jogo -> Jogada -> Jogador
posjogador (Jogo (Jogador (x,y)) (Mapa largura linhas)) Parado = Jogador (x,y)
posjogador (Jogo (Jogador (x,y)) (Mapa largura linhas)) (Move Cima) = if y == 0  || ((snd(linhas !! (y-1))) !! x) == Arvore
                                                                      then Jogador (x,y)
                                                                      else Jogador (x,y-1)

posjogador (Jogo (Jogador (x,y)) (Mapa largura linhas)) (Move Baixo) = if (y) == ((length linhas)-1) || ((snd(linhas !! (y+1))) !! x) == Arvore
                                                                       then Jogador (x,y) 
                                                                       else Jogador (x,y + 1) 

posjogador (Jogo (Jogador (x,y)) (Mapa largura linhas)) (Move Esquerda) = if x == 0 || ((snd(linhas !! (y))) !! (x-1)) == Arvore
                                                                          then Jogador (x,y) 
                                                                          else Jogador (x - 1,y)  

posjogador (Jogo (Jogador (x,y)) (Mapa largura linhas)) (Move Direita) = if x == (largura -1) || ((snd(linhas !! (y))) !! (x+1)) == Arvore
                                                                         then Jogador (x,y) 
                                                                         else Jogador (x + 1,y)

Propriedades:

O jogador só pode sair do mapa através do movimento de um dos troncos.

O jogador não poderá subir a nenhuma Arvore.

Função localizajogador

localizajogador :: Jogo -> [Obstaculo]

A função localizajogador recebe um Jogo e dá como resultado uma lista do tipo Obstaculo. O objetivo desta função é encontrar a lista de objetos da linhas onde o jogador se encontra.

localizajogador :: Jogo -> [Obstaculo]
localizajogador (Jogo (Jogador (x,y)) (Mapa largura linhas)) = (snd (linhas !! y)) 

Exemplo:

>>> localizajogador (Jogo (Jogador (0,1)) (Mapa 2 [(Rio 1, [Nenhum,Tronco]),(Estrada 2, [Nenhum,Carro])])) = [Nenhum,Carro]

Função localizavelocidade

localizavelocidade :: Jogo -> Velocidade

A função localizavelocidade recebe um Jogo e dá como resultado uma Velocidade. O objetivo desta função é encontrar a Velocidade do Rio onde o jogador se encontra.

localizavelocidade :: Jogo -> Velocidade 
localizavelocidade (Jogo (Jogador (x,y)) (Mapa largura linhas)) = veloaux (fst (linhas !! y))
    where veloaux (Rio v) = v  

Exemplo:

>>> localizavelocidade (Jogo (Jogador (0,1)) (Mapa 2 [(Rio 1, [Nenhum,Tronco]),(Rio 2, [Nenhum,Carro])])) = 2