Tarefa5_2022li1g010

Description

Este módulo contem a resolução da Tarefa 5, a qual promove o efeito de deslize do mapa.

Função deslizaJogo

deslizaJogo :: Int -> Jogo -> Jogo

A função deslizaJogo recebe um Int e um Jogo e dá como resultado um Jogo. O objetivo desta função é efetuar o deslize do mapa de acordo com a geração aleatória implementada na Tarefa 2.

localizajogador :: Jogo -> [Obstaculo]
localizajogador (Jogo (Jogador (x,y)) (Mapa largura linhas)) = (snd (linhas !! y))